requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
cftime==1.6.2
charset-normalizer==3.3.0
Cython==3.0.3
entsoe-py==0.5.10
holidays==0.34
ipykernel==6.25.2
jupyter_client==8.4.0
jupyter_core==5.4.0
linearmodels==5.3
matplotlib==3.8.0
netCDF4==1.6.4
networkx==3.3
numpy==1.23.4
openpyxl==3.1.2
pandas==2.1.1
pmdarima==2.0.4
plotly==5.17.0
scikit-learn==1.5.0
scipy==1.11.3
statsmodels==0.14.0
suntime==1.2.5
